requirements.txt
requests==2.22.0
pandas==0.25.3
selenium==3.141.0
numpy==1.19.5
bs4==0.0.1
lxml==4.6.2
python-Levenshtein==0.12.0
dlib==19.21.1
face-recognition==1.3.0
opencv-python==4.4.0.46
rawpy==0.16.0